// src/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Decimal: any;
};
// ...
export type LoginHistory = {
  __typename?: 'LoginHistory';
  accessDevice?: Maybe<Scalars['String']>;
  accessLocation?: Maybe<Scalars['String']>;
  browser?: Maybe<Scalars['String']>;
  createdAt?: Maybe<Scalars['DateTime']>;
  ipAddress?: Maybe<Scalars['String']>;
};
// ...
export type User = {
  __typename?: 'User';
  dob?: Maybe<Scalars['String']>;
  email: Scalars['String'];
  firstName?: Maybe<Scalars['String']>;
  gender?: Maybe<Scalars['String']>;
  google2faSecret?: Maybe<Scalars['String']>;
  image?: Maybe<Scalars['String']>;
  is2faEnabled?: Maybe<Scalars['Boolean']>;
  isEmailVerified?: Maybe<Scalars['Boolean']>;
  isWhitelistEnabled?: Maybe<Scalars['Boolean']>;
  lastName?: Maybe<Scalars['String']>;
  loginHistories?: Maybe<Array<Maybe<LoginHistory>>>;
  mobileNumber?: Maybe<Scalars['String']>;
  phoneNumber?: Maybe<Scalars['String']>;
  receiveNewsletter?: Maybe<Scalars['Boolean']>;
  referenceCode?: Maybe<Scalars['String']>;
  userAddress?: Maybe<Array<Maybe<UserAddress>>>;
  userId: Scalars['String'];
  userName?: Maybe<Scalars['String']>;
  userRole?: Maybe<Array<Maybe<UserRole>>>;
};

export type UserAddress = {
  __typename?: 'UserAddress';
  City?: Maybe<Scalars['String']>;
  State?: Maybe<Scalars['String']>;
  Zipcode?: Maybe<Scalars['String']>;
  addressLine1?: Maybe<Scalars['String']>;
  addressLine2?: Maybe<Scalars['String']>;
};

export type UserMembership = {
  __typename?: 'UserMembership';
  isCurrent: Scalars['Boolean'];
  membershipEndedDate?: Maybe<Scalars['DateTime']>;
  membershipProgramId: Scalars['String'];
  membershipPrograms: MembershipProgram;
  membershipStartDate: Scalars['DateTime'];
  userId: Scalars['String'];
};

export type UserRole = {
  __typename?: 'UserRole';
  userId: Scalars['String'];
  userRoleTypeId?: Maybe<Scalars['Int']>;
};